requirements.txt
pandas
plotly
streamlit
numpy
requests
Pillow
